Move pkg/finder -> pkg/processor/finder
