+++ internal/pkg/processor/finder/fixtures/complex/files/foo1.bar
