+++ internal/pkg/processor/finder/fixtures/complex/files/foo2.bar
