+++ internal/pkg/processor/finder/fixtures/complex/files/foo3.csv
