+++ internal/pkg/processor/finder/fixtures/complex/files/sub/dir/foo4.bar
