+++ internal/pkg/processor/finder/fixtures/complex/files/sub/dir/foo5.bar
