+++ internal/pkg/processor/finder/fixtures/complex/files/sub/dir/foo6.csv
