+++ internal/pkg/processor/finder/fixtures/complex/tables/sub/dir/foo7.bar
