+++ internal/pkg/processor/finder/fixtures/complex/tables/sub/dir/foo8.csv
