+++ internal/pkg/processor/finder/fixtures/complex/tables/table1.csv
