+++ internal/pkg/processor/finder/fixtures/complex/tables/table2.csv.manifest
